Eliminated delay() in MakeStrings file and cleaned up formatting
MakeStrings code talks to magnetometer and makes a string for the
magnetometer data every 1000ms. The time interval can be changed easily.
I added the magnetometer example code for understanding the I2C syntax
with using the magnetometer, and the trivial switch test code for an
example of using the pushbutton switch for starting a process